Add stack guard for v8086
Some BIOS handlers decide to far ret back to the caller instead of simply using
the iret instruction. Since we don't intercept ret instructions in
v8086_gpf_handler, this would cause the stack to underflow.

This patch adds a very small (5 byte) stack guard which simply contains the
(far) address of the next byte, followed by the iret opcode. That makes sure
that the final iret is always called.

Correspondingly, the GP handler now has to accept that the stack might have
already underflowed, and return to protected mode in any case.