Fix return address in guard
The guard return address was pointing at the (zero) instruction before the
intended iret, i.e. add %al, (%eax). That caused the carry flag to be set when
calling a BIOS interrupt and it appeared as if the BIOS routine was signalling
failure.

I found the issue by adding some tracing and observing the stack just before
the final return.