Use a special trampoline for scheduling
My previous plan was to just jump to a task using the existing iret in
isr_generic, but that one does not restore the stack if we are preempting a
kernel-mode task. Now I'm using a very small trampoline which assumes that the
saved stack is at the beginning of the isr_stack_t structure.