Always keep preemption disabled in a handler
Because of the way the current context switching works, if a task gets
preempted while in an interrupt context, it will wake up outside the handler.
So for example if the handler masked the corresponding IRQ, it will never be
unmasked.

This was causing the keyboard IRQ handler to stop working. I fixed it by
keeping interrupts disabled there. It could probably be improved by reenabling
IRQs but keeping preemption disabled, but that is slightly harder to achieve,
because the premption functions assume to be run outside an interrupt context.