Improve console event loop
The console is not on a timer anymore. It redraws whenever it needs to.
There are two semaphores: a global console lock write_sem, and a display
semaphore paint_sem. To write to the console, a task takes the lock, writes,
then signals to the console that a redraw is needed. The renderer task waits
on the paint semaphore in a loop, and keeps redrawing as necessary.