Fix ISR stack size
The isr_stack_t structure was larger than necessary, to accomodate for a
larger stack frame on a context switch interrupt (like when going from v8086
back to kernel mode).

This caused a problem when setting up the stack guard for a task, since the
return address instruction was not placed on the correct position in the
stack.

This commit fixes the problem by introducing a new struct v8086_isr_stack_t
to hold the stack frame for a v8086 task switch, while using the previous
struct for a generic interrupt handler.